+++ job_runner.py
@@ -20,6 +20,7 @@ from app.email_utils import (
     render,
     get_email_domain_part,
 )
+from app.import_utils import handle_batch_import
 from app.utils import sanitize_email
 from app.extensions import db
 from app.log import LOG
@@ -111,84 +112,6 @@ def onboarding_mailbox(user):
         unsubscribe_link,
         via_email,
     )
-
-
-def handle_batch_import(batch_import: BatchImport):
-    user = batch_import.user
-
-    batch_import.processed = True
-    db.session.commit()
-
-    LOG.debug("Start batch import for %s %s", batch_import, user)
-    file_url = s3.get_url(batch_import.file.path)
-
-    LOG.d("Download file %s from %s", batch_import.file, file_url)
-    r = requests.get(file_url)
-    lines = [line.decode() for line in r.iter_lines()]
-    reader = csv.DictReader(lines)
-
-    for row in reader:
-        try:
-            full_alias = sanitize_email(row["alias"])
-            note = row["note"]
-        except KeyError:
-            LOG.warning("Cannot parse row %s", row)
-            continue
-
-        alias_domain = get_email_domain_part(full_alias)
-        custom_domain = CustomDomain.get_by(domain=alias_domain)
-
-        if (
-            not custom_domain
-            or not custom_domain.verified
-            or custom_domain.user_id != user.id
-        ):
-            LOG.debug("domain %s can't be used %s", alias_domain, user)
-            continue
-
-        if (
-            Alias.get_by(email=full_alias)
-            or DeletedAlias.get_by(email=full_alias)
-            or DomainDeletedAlias.get_by(email=full_alias)
-        ):
-            LOG.d("alias already used %s", full_alias)
-            continue
-
-        mailboxes = []
-
-        if "mailboxes" in row:
-            for mailbox_email in row["mailboxes"].split():
-                mailbox_email = sanitize_email(mailbox_email)
-                mailbox = Mailbox.get_by(email=mailbox_email)
-
-                if not mailbox or not mailbox.verified or mailbox.user_id != user.id:
-                    LOG.d("mailbox %s can't be used %s", mailbox, user)
-                    continue
-
-                mailboxes.append(mailbox.id)
-
-        if len(mailboxes) == 0:
-            mailboxes = [user.default_mailbox_id]
-
-        alias = Alias.create(
-            user_id=user.id,
-            email=full_alias,
-            note=note,
-            mailbox_id=mailboxes[0],
-            custom_domain_id=custom_domain.id,
-            batch_import_id=batch_import.id,
-        )
-        db.session.commit()
-        db.session.flush()
-        LOG.d("Create %s", alias)
-
-        for i in range(1, len(mailboxes)):
-            alias_mailbox = AliasMailbox.create(
-                alias_id=alias.id,
-                mailbox_id=mailboxes[i],
-            )
-            db.session.commit()
-            LOG.d("Create %s", alias_mailbox)
 
 
 if __name__ == "__main__":
